requirements.txt
annotated-types==0.5.0
ansicolors==1.1.8
arango==0.2.1
argparse-color-formatter==1.2.2.post2
async-timeout==4.0.3
azure-common==1.1.28
azure-core==1.29.4
azure-cosmos==4.5.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
basho-erlastic==2.1.1
bcrypt==4.0.1
black==23.7.0
bleach==6.0.0
boto3==1.28.25
botocore==1.31.25
build==0.10.0
cachetools==5.3.1
cassandra-driver==3.28.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
couchbase==4.1.8
coverage==7.3.0
cryptography==41.0.3
cx-Oracle==8.3.0
direnv==2020.12.3
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
elastic-transport==8.4.0
elasticsearch==8.9.0
emoji==2.7.0
env-file==2020.12.3
exceptiongroup==1.1.2
flake8==6.1.0
geniusrise==0.0.3
geomet==0.2.1.post1
google-api-core==2.11.1
google-auth==2.17.3
google-cloud-bigquery==3.11.4
google-cloud-bigtable==2.21.0
google-cloud-core==2.3.3
google-cloud-spanner==3.40.1
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
happybase==1.2.0
hdfs==2.7.2
ibm-db==3.2.0
idna==3.4
importlib-metadata==6.8.0
influxdb==5.3.1
iniconfig==2.0.0
ipaddress==1.0.23
jaraco.classes==3.3.0
jeepney==0.8.0
jmespath==0.10.0
jsonpickle==3.0.1
kafka-python==2.0.2
keyring==24.2.0
kubernetes==27.2.0
ldap3==2.9.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
memsql==3.2.0
mock==5.1.0
more-itertools==10.1.0
msgpack==1.0.5
mypy==1.5.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
neo4j==5.12.0
nose==1.3.7
oauthlib==3.2.2
opentsdb-py==0.6.0
packaging==23.1
paramiko==3.3.1
pathspec==0.11.2
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.2.0
ply==3.11
presto-python-client==0.8.4
prettytable==3.8.0
prometheus-client==0.17.1
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
pymssql==2.2.8
PyMySQL==1.1.0
PyNaCl==1.5.0
pynuodb==2.6.0
pyodbc==4.0.39
pyproject_hooks==1.0.0
pytest==7.4.0
python-arango==7.6.2
python-dateutil==2.8.2
python-ldap==3.4.3
python3-protobuf==2.5.0
pytz==2023.3.post1
PyYAML==6.0.1
readme-renderer==40.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==2.0.0
riak==2.7.0
rich==13.5.2
rich-argparse==1.3.0
rsa==4.9
s3transfer==0.6.1
SecretStorage==3.3.3
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
termcolor==2.3.0
thriftpy2==0.4.16
tomli==2.0.1
twine==4.0.2
types-PyMySQL==1.1.0.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.6
typing_extensions==4.7.1
urllib3==1.26.16
values==2020.12.3
vertica-python==1.3.5
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
Wraptor==0.7.0
zipp==3.16.2
